Consolidate some tests
test_md5sum is more of a server test than anyting else.
