+++ lib/pbench/test/server/test_utils.py
@@ -1,6 +1,9 @@
-import pytest
-from pbench.server.utils import filesize_bytes
+from pathlib import Path
 
+import pytest
+
+from pbench.common.utils import md5sum
+from pbench.server.utils import filesize_bytes
 
 _sizes = [("  10  ", 10)]
 for i, mult in [
@@ -34,3 +37,16 @@ class TestFilesizeBytes:
             ), f"string '{size}', improperly converted to {res}, expected {exp_val}"
         with pytest.raises(Exception):
             res = filesize_bytes("bad")
+
+
+class TestMd5sum:
+    @staticmethod
+    def test_md5sum():
+        # Reuse the existing file from the server upload fixture.
+        filename = "log.tar.xz"
+        test_file = Path("./lib/pbench/test/server/fixtures/upload/", filename)
+        expected_hash_md5 = open(f"{test_file}.md5", "r").read().split()[0]
+        hash_md5 = md5sum(test_file)
+        assert (
+            hash_md5 == expected_hash_md5
+        ), f"Expected MD5 '{expected_hash_md5}', got '{hash_md5}'"
